LC3 code for compsci210
Holds the .asm files for the LC3 (Little computer 3) This code takes
input from the user, then it prints out the input how every many times
the user specifies in this case age. In the code is a conversion method
that takes the ASCII values and makes them into decimal equivalents.